tx/types: move batch parsing logic into types header
Break the monolith of rm_stm further and move unnecessarly logic
out of it.
